+++ cmd/install/install.go
@@ -1,0 +1,14 @@
+package install
+
+import (
+	"github.com/raxigan/pcfy-my-mac/cmd/param"
+	"time"
+)
+
+type Installation struct {
+	Commander
+	param.Params
+	HomeDir
+	ProfileName      string
+	InstallationTime time.Time
+}
