+++ test/params_test.go
@@ -1,7 +1,8 @@
 package install_test
 
 import (
-	"github.com/raxigan/pcfy-my-mac/install"
+	"github.com/raxigan/pcfy-my-mac/cmd/common"
+	"github.com/raxigan/pcfy-my-mac/cmd/param"
 	"github.com/raxigan/pcfy-my-mac/test/test_utils"
 	"testing"
 )
@@ -9,7 +10,7 @@ import (
 func TestFailForUnknownParam(t *testing.T) {
 
 	yml := test_utils.Trim(`unknown: hello`)
-	_, err := install.CollectYamlParams(yml)
+	_, err := param.CollectYamlParams(yml)
 
 	test_utils.AssertErrorContains(t, err, "Unknown parameter: unknown")
 }
@@ -17,14 +18,14 @@ func TestFailForUnknownParam(t *testing.T) {
 func TestFailForInvalidYaml(t *testing.T) {
 
 	yml := test_utils.Trim(`[] :app-launcher:`)
-	_, err := install.CollectYamlParams(yml)
+	_, err := param.CollectYamlParams(yml)
 
-	test_utils.AssertErrorContains(t, err, "cannot unmarshal !!seq into install.FileParams")
+	test_utils.AssertErrorContains(t, err, "cannot unmarshal !!seq into param.FileParams")
 }
 
 func TestInstallYmlFileDoesNotExist(t *testing.T) {
 
-	_, err := install.TextFromFile("i-do-not-exist.yml")
+	_, err := common.TextFromFile("i-do-not-exist.yml")
 
 	test_utils.AssertErrorContains(t, err, "open i-do-not-exist.yml: no such file or directory")
 }
@@ -32,7 +33,7 @@ func TestInstallYmlFileDoesNotExist(t *testing.T) {
 func TestInstallInvalidAppLauncher(t *testing.T) {
 
 	yml := test_utils.Trim(`app-launcher: unknown`)
-	_, err := install.CollectYamlParams(yml)
+	_, err := param.CollectYamlParams(yml)
 
 	test_utils.AssertErrorContains(t, err, `Invalid param 'app-launcher' value/s 'unknown', valid values:
 		spotlight
@@ -44,7 +45,7 @@ func TestInstallInvalidAppLauncher(t *testing.T) {
 func TestInstallInvalidTerminal(t *testing.T) {
 
 	yml := test_utils.Trim(`terminal: unknown`)
-	_, err := install.CollectYamlParams(yml)
+	_, err := param.CollectYamlParams(yml)
 
 	test_utils.AssertErrorContains(t, err, `Invalid param 'terminal' value/s 'unknown', valid values:
 		default
@@ -56,7 +57,7 @@ func TestInstallInvalidTerminal(t *testing.T) {
 func TestInstallInvalidKeyboardLayout(t *testing.T) {
 
 	yml := test_utils.Trim(`keyboard-layout: unknown`)
-	_, err := install.CollectYamlParams(yml)
+	_, err := param.CollectYamlParams(yml)
 
 	test_utils.AssertErrorContains(t, err, `Invalid param 'keyboard-layout' value/s 'unknown', valid values:
 		pc
